Tweaked database indexes
`rhymes_with` now also contains `prerhyme` so that rhyming joins can be
convering. `notions_lemmas` and `lemmas_notions` have been created so as
to faciliate "jumping" over `words` when it's only needed as a
many-to-many through table. Because `notion_words` and `lemma_words` are
prefixes of these new indexes, they have been removed.